% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDGCopula.R
\name{cosineSim}
\alias{cosineSim}
\title{Compare angles}
\usage{
cosineSim(orig, syn)
}
\arguments{
\item{orig}{original data.}

\item{syn}{synthetic data}
}
\value{
value between -1 to 1 which indicate similarity between two data frames.
}
\description{
Compare angles
}
\examples{
cosineSim(orig, syn)
}
