% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDGCopula.R
\name{synthesize}
\alias{synthesize}
\title{This function returns a SDGCopula object.}
\usage{
synthesize(dataframe, copula, parametric, dof)
}
\arguments{
\item{dataframe}{from which synthetic data is to be generated.}

\item{copula}{normal or t, default = 'normal'}

\item{parametric}{TRUE or FALSE, default = FALSE}

\item{dof}{degree of freedom for t copula, default dof = 4}
}
\value{
SDGCopula object containing synthetic dataframe, RMSE and Cosine Similarity.
}
\description{
This function returns a SDGCopula object.
}
\examples{
SDGCopula(dataframe = df,copula = normal, parametric = F)
}
