% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDGCopula.R
\name{histfit}
\alias{histfit}
\title{A function to compare the histograms of original and synthetic data.}
\usage{
histfit(df, position, ylim)
}
\arguments{
\item{df}{Data frame containing both original and synthetic data.}

\item{position}{Position of the legend.}

\item{ylim}{Set y-axis limits if necessary.}
}
\value{
Histogram plot comparing original and synthetic data.
}
\description{
A function to compare the histograms of original and synthetic data.
}
\examples{
histfit(df = dataframe, position = 'topright')
}
