% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDGCopula.R
\name{marginals}
\alias{marginals}
\title{Function to obtain the parametric estimates of marginals.}
\usage{
marginals(dataframe)
}
\arguments{
\item{dataframe}{}
}
\description{
Function to obtain the parametric estimates of marginals.
}
