% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDGCopula.R
\name{RMSE}
\alias{RMSE}
\title{Function to calculate root mean square error.}
\usage{
RMSE(orig, syn)
}
\arguments{
\item{orig}{original data.}

\item{syn}{synthetic data.}
}
\value{
RMSE value.
}
\description{
Function to calculate root mean square error.
}
\examples{
RMSE(orig=data,syn=synthetic_data)
}
