% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDGCopula.R
\name{tocategory}
\alias{tocategory}
\title{Function to convert numeric values to the original categorical variables.}
\usage{
tocategory(x)
}
\arguments{
\item{x}{}
}
\description{
Function to convert numeric values to the original categorical variables.
}
